PycharmProjects/ChatChatChat/ChatChat/requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.9
django-cors-headers==4.0.0
django-rest-swagger==2.2.0
django-summernote==0.8.20.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
webencodings==0.5.1
